# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing and querying time series data from IoT sensors\n\nThis API allows you to upload CSV files containing time series data,\nquery the data with time range filters, and manage datasources.\n\nSupported timestamp formats: ISO8601, RFC3339, Unix timestamps (seconds/milliseconds), Julian Day",
        "title": "IoT Data Sandbox API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/datasources": {
            "get": {
                "description": "Get a list of all registered datasources with their metadata",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasources"
                ],
                "summary": "List all datasources",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.DataSourceListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Upload a CSV file containing time series data. The CSV must have 'timestamp' and 'value' columns. Supports various timestamp formats (ISO8601, Unix, Julian Day).",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasources"
                ],
                "summary": "Upload a CSV datasource",
                "parameters": [
                    {
                        "type": "file",
                        "description": "CSV file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name for the datasource (defaults to filename)",
                        "name": "name",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/data.UploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/datasources/{id}": {
            "get": {
                "description": "Get metadata for a specific datasource by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasources"
                ],
                "summary": "Get datasource metadata",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Datasource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.DataSourceMetadata"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a datasource and its associated CSV file",
                "tags": [
                    "datasources"
                ],
                "summary": "Delete a datasource",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Datasource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/datasources/{id}/data": {
            "get": {
                "description": "Query time series data from a datasource with optional time range filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasources"
                ],
                "summary": "Query time series data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Datasource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start time in RFC3339 format (e.g., 2024-01-01T00:00:00Z)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time in RFC3339 format (e.g., 2024-01-01T23:59:59Z)",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.DataQueryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/layers/{id}": {
            "get": {
                "description": "Get a specific layer by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "layers"
                ],
                "summary": "Get layer details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Layer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.LayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a layer by ID",
                "tags": [
                    "layers"
                ],
                "summary": "Delete a layer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Layer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/layers/{id}/color": {
            "put": {
                "description": "Update the display color of a layer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "layers"
                ],
                "summary": "Update layer color",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Layer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Color value (hex format)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.UpdateColorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.LayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/layers/{id}/data": {
            "get": {
                "description": "Get the time series data points for a layer (includes data from associated datasource)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "layers"
                ],
                "summary": "Get layer time series data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Layer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Start time in RFC3339 format (e.g., 2024-01-01T00:00:00Z)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time in RFC3339 format (e.g., 2024-01-01T23:59:59Z)",
                        "name": "end_time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.DataQueryResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/layers/{id}/display-window": {
            "put": {
                "description": "Update the display time window (start and end times) for a layer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "layers"
                ],
                "summary": "Update layer display time window",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Layer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Display window times",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.UpdateDisplayWindowRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.LayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/layers/{id}/duplicate": {
            "post": {
                "description": "Create a copy of a layer with a new name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "layers"
                ],
                "summary": "Duplicate a layer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Layer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New layer name",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.DuplicateLayerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/data.LayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/layers/{id}/load-csv": {
            "post": {
                "description": "Load data from an existing CSV file into the layer (creates a new datasource)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "layers"
                ],
                "summary": "Load CSV data into layer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Layer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "CSV filename",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.LoadCSVRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.LayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/layers/{id}/visibility": {
            "put": {
                "description": "Update the visibility state of a layer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "layers"
                ],
                "summary": "Update layer visibility",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Layer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Visibility state",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.UpdateVisibilityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.LayerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/projects": {
            "get": {
                "description": "Get a list of all projects",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "List all projects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.ProjectListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new empty project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Create a new project",
                "parameters": [
                    {
                        "description": "Project details",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.CreateProjectRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/data.ProjectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "get": {
                "description": "Get a specific project by ID with layer count",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projects"
                ],
                "summary": "Get project details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/data.ProjectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a project and all its layers (cascade)",
                "tags": [
                    "projects"
                ],
                "summary": "Delete a project",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/data.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "data.DataQueryResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.DataPoint"
                    }
                },
                "end_time": {
                    "type": "string"
                },
                "row_count": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "data.DataSourceListResponse": {
            "type": "object",
            "properties": {
                "data_sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.DataSourceMetadata"
                    }
                }
            }
        },
        "data.DataSourceMetadata": {
            "type": "object",
            "properties": {
                "data_source_id": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "row_count": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "time_label": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value_label": {
                    "type": "string"
                },
                "when_created": {
                    "type": "string"
                }
            }
        },
...
        "data.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "data.UpdateColorRequest": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                }
            }
        },
        "data.UpdateDisplayWindowRequest": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
...
        "data.UploadResponse": {
            "type": "object",
            "properties": {
                "data_source_id": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "row_count": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "time_label": {
                    "type": "string"
                },
                "value_label": {
                    "type": "string"
                },
                "when_created": {
                    "type": "string"
                }
            }
        }
    }
}